+++ Game/items/weapon/MagicTorch.java
@@ -2,7 +2,7 @@ package items.weapon;
 
 public class MagicTorch extends Weapon{
 
-    protected MagicTorch(String name, float weigth, int damage, int streigth) {
+    protected MagicTorch(String name, int weigth, int damage, int streigth) {
         super(name, weigth, damage, streigth);
     }
