+++ Game/items/weapon/Rifle.java
@@ -2,7 +2,7 @@ package items.weapon;
 
 public class Rifle extends Weapon{
 
-    protected Rifle(String name, float weigth, int damage, int streigth) {
+    protected Rifle(String name, int weigth, int damage, int streigth) {
         super(name, weigth, damage, streigth);
     }
