+++ Game/items/weapon/Sword.java
@@ -2,7 +2,7 @@ package items.weapon;
 
 public class Sword extends Weapon{
 
-    public Sword(String name, float weigth, int damage, int streigth) {
+    public Sword(String name, int weigth, int damage, int streigth) {
         super(name, weigth, damage, streigth);
     }
